Major changes to class interfaces and naming conventions
Changed many of the class methods and removed some that weren't necessary.

Changed naming convention on string literals to all capital letters.

Created an events class for moving through the game event loop.

More work is needed on the game class and separating the classes as well as the logic that handles cell actions and game events.

Changes are required to some of the clone methods and the serialize/deserialze functions